+++ src/qt_controls_pyrs/buttons/__init__.py
@@ -2,5 +2,17 @@
 
 from .framebutton import FrameButton
 from .glowbutton import GenerateButton, GlowButton
+from .hoverbuttons import (
+    DashBorderButton, HaloButton, RaisedButton, ShineButton, SpreadButton
+)
 
-__all__ = ["FrameButton", "GenerateButton", "GlowButton"]
+__all__ = [
+    "DashBorderButton",
+    "FrameButton",
+    "GenerateButton",
+    "GlowButton",
+    "HaloButton",
+    "RaisedButton",
+    "ShineButton",
+    "SpreadButton",
+]